requirements.txt
tensorflow-gpu<2.0 # or tensorflow==1.15
matplotlib
scipy
hickle<4.0
torch
tensorflow_probability<0.9
torchdiffeq
